Fix node redraw when escaping from draw mode
This is the case where the Array of changed element IDs
contains an ID for which there is no corresponding entity
in the graph. While `all` will not contain this entity,
the filter should still return true given its id, so that
it will appear in the exit selection and be removed.