Initial stab at an ImproveWay mode
This mode helps you to align a way to aerial imagery, or
to draw more of an existing way. On mouse events, it either
adds a new node at the current mouse position between the
two vertices that are the endpoints of the closest edge,
or moves an existing vertex. The heuristic for choosing
between those two needs some tweaking. The main problem
currently is that the vertex wants to move only on the
reflex (>180°) side of the angle, and not the other.