Ensure combobox suggestion menu is closed on blur
The combobox's blur handler needs to be called prior
to the tag editor's blur handler. The latter can cause
a re-render, which will rebind handlers. Rebound handlers
are not called during the same event dispatch cycle, so
when the combobox blur handler is last, it sometimes
doesn't get called, stranding the menu.