Fix member list redraws
d3 does not store the keys produced by a data bind -- it
recomputes them on the fly. The key function that was used
referred to `entity`, with the intent that rows referring
to the previous entity would wind up in the exit selection.
But since the key is recomputed for the existing bindings,
and at that point `entity` refers to the new entity, that
portion of the key effectively always matched.

While here, fix deletion of incomplete members (don't know
what the geometry is, so can't do geometry-specific annotations).

Fixes #1548.