remove addr:housename from presets
The [values](http://taginfo.openstreetmap.org/keys/addr%3Ahousename#values) of this rarely needed tag, which is prominent in the address presets, now have a lot of numbers and descriptive words in many different languages. Removing the preset means that new editors won’t fill this key with information useless to the key but can still add it through All tags if needed.  A similar thing happened when Potlatch put the designation tag into presets. Goes towards #1427 and #2124.