Do not fully fill certain landuse values, e.g. landuse=residential
residential, retail, commercial and industrial landuse tags
typically cover large, dense areas with many overlapping
features. The large click targets of filled areas have proven
to be very good for usability in the general case, but are
a common source of accidental mistagging in this case.

We may add additional tags to this list, but these are the
obvious first step — especially landuse=residential.

Fixes #542.