map#trimmedExtentZoom
Like map#extentZoom, but uses trimmed viewport instead of full viewport
to avoid putting the extent under UI buttons and border elements.

This is because in conflict resolution the user will be extentZooming to see
their changes and it was annoying to have the change appear under the UI.

Also using this in zoomToGPXLayer.