Add `snake_case` field property (closes #2867)
The vast majority of tag values in OSM should be automatically snake_cased
(e.g. `amenity=place_of_worship`)
But there are a few exceptions, e.g. fields expected to include a unit
(e.g. `maxstay=1 day`)
So this allows a ui field to set whether or not to automatically snake_case