Add fire-station icon for fire-hydrant
Fire hydrants are for firemen anyway, and it help distinguish them from other no-icon point on the map. 

By the way, is there any plan to upgrade maki from version 0.5.0 to latest version 2.0.x? Is that OK if I migrate iD by updating icons' name or there is a particular reason to use maki 0.5.0?