Updated http links to https
I updated http links in the readme that would otherwise automatically be redirected to https by the website, to just use https directly. :)

Thank you again for making an awesome editor! ^ _ ^

I'd like to note that links in this section of the readme
"
Test a prerelease version of iD:
Stable mirror of release branch: http://openstreetmap.us/iD/release
Development mirror of master branch: http://openstreetmap.us/iD/master
"

Currently 404. I'm not sure if this is temporary or permanent.