Update column.json
An ad-column is a round thing the way I know it and the way it is shown in the wiki (https://wiki.openstreetmap.org/wiki/Tag:advertising%3Dcolumn). So the direction key as preset does not make sense. There seems to be no relevant usage of this combination as well (https://taginfo.openstreetmap.org/tags/advertising=column#combinations).