Remove setCenter() from map.dimensions()
This was triggering a weird redraw which made the map slow and look wrong
The map will redraw again in a few milleseconds anyway.

This means that the the map will not stay on the center, but rather the
top-left corner, which kind of feels better anyway, as it doesn't move things
around as much.