Switch brand suggestion preset generation over to inherit fields
This also attempts to use `cuisine` and `vending` tags in the
presetID - in case a more specific preset exists for those things.

This fixes a minor annoyance where only matching `key/value`
would find the "fast food" preset instead of the "chicken fast food"
or "pizza fast food" preset.

So now Chick-fil-a has a chicken icon and Sbarro has a pizza icon. :pizza: