Fix one-way error positioning and description
The direction of travel in the description was misleading as the way
segment wasn't necessary linear and so the direction could change as you
travel along the way. This is more explicit as it specifies the detected
start/end nodes (also useful to show that it's not the whole way that
might be one-way).

The position is now taken as the central node in the `feature.points`
list (or the average of the central two when there are an even number of
points).