remove building=shed from bicycle_parking=shed preset
building=shed should not be added if the geometry is a point, but should, if possible, if it is an area. Unfortunately, the iD preset system is currently not able to map this behavior to a preset. Instead, there is the building_area field which automatically adds building=yes to the preset if it is an area.