Favorite backgrounds with next/previous keybinding
Binds , to previous background in list, and . to next background in
list. The thinking to my mind is that those are under the < and >
characters on a US keyboard, but that may not be international enough.
This takes the current list of backgrounds, in displayed order and
allows you to iterate through it forwards or backwards. It will only do
the custom background if there's a custom background defined.

This also introduces the concept of favorite backgrounds, which come
together at the top of the background list, for easy grouping together
to use the next/previous keys on. It also shifts the best nomenclature
to use a different graphic from the star, as that's now meaning
favorite.

One open question for me is the addition of the keybinding work inside
background_list.js, rather than where it seems like most of them are
occuring (in init.js). If we wanted we could do some heavy lifting to
get the appropriate stuff surfaced within init, but I like this
architecture as it has the keybindings close to the functionality they
link to.

A candidate area to navigate around in to see the functionality is:
http://127.0.0.1:8080/#map=16.00/51.0000/4.0000