Add ability to snap a way onto another way
Fixes #9223

Added an action that allows to snap 2 ways together.
Created the associated operation for this action.
Added some fonctions that are used by the snap action in the way.js file.
Snap allows to speed up the follow process as it is possible to snap the 2 ways together with only one click.

Also added unit tests for this feature.