use addr+place subfield for Taiwan
this assumes that the address format included both fields as a "hacky" workaround of iD's inability to choose between streets and places from a single field. This is (at the momemt) only based on the observation that before the format was introduced in iD in mid-2016, there were only a very small number of addresses in Taiwan with both tags, and the number increased only after iD started to show fields for both tags, and even today only a very small percentage of addresses have both tags set (approx 2.1k out of 1.6M).