Add support for a tilejson vector layer.
Prefer using the content type rather than the filename extension to
decide the format of a custom data file.

Should probably specifically ignore .txt and .bin which indicate a
confused web server or web browser.

Should verify that web browsers supply something useful for the file's
type on various platforms, or else change this back to checking the file
extension for local files.