suppress autocomplete for taginfo results if field has static values
fixes #9898

Fixes a problem that occurs because the order of the field `options`
can be in general different from the order returned by taginfo
(taginfo sorts by usage, while presets typically are sorted alphabetically),
causing different values to be potentially selected for autocomplete values.
As the taginfo results typically come in a few moments delayed, this can
lead to confusing behaviour.

This solves it by just not reattempting an autocomplete using the taginfo
results, if a field has static options available.