feat(fields/combo): add navigate button to web links in values
The common use case is for the `sources` key, often you have multiple URLs separated with semicolon.
This allow to navigate to URL and check them more easily instead of
doing manual work to select the URL copy and paste.
Created with the help of Martin Raifer at the sotmeu2023!

Signed-off-by: Matthieu Bessat <mail@matthieubessat.fr>