fix zoom/pan glitch on plane photos
When changing to a different photo, it can be the case that the new photo has a different aspect ratio: the zoom behaviour needs to be updated.

Also, keeping the "orientation" does not make sense for plane photos, and even keeping the previously shown photo's zoom is also not typically desired for these.