fix failing tests
location-conflation v2 uses a Map instead of an object for its internal cache, which needed to be adjusted accordingly.

It would be ideal to not having to rely on undocumented internal properties of the library, but unfortunately, it does not provide the required functionality in the public API at the moment.